Rework authentication
We no longer attach a full Author object (which required a database
query with each request) to the context when authenticated. A new
Authentication type is added which is always present on the context. The
Authenticatoin type's `loggedIn` property can be used to check whether
the user is logged in. If a used is logged in, the `id` property refers
to their ID, which is pulled straight from the JWT presented with the
request.

Resolves the issue of models being created before authentication had
completed, resulting in the destructuring of the author object being
undefined within the model generator function context argument.