Renamed Post to BlogPost, PostEdit to Edit
Added Content interface as a container for a title and content, which
BlogPost implements. Will be used later for JournalEntry as well. Edit
will be used for both BlogPost and JournalEntry.

Title and content now stored in 'content' table which is joined in
queries.

Use transactions for operations which require multiple queries to
complete (creating a new author or post, updating a post)

Delete post IDs from uuids table to cause cascding delete.